+++ management/up.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+set -euo pipefail
+
+terraform init
+terraform apply -auto-approve