+++ CHANGELOG.md
@@ -5,5 +5,7 @@
   - Start with Terraform `.gitignore`
   - DevContainer/CodeSpaces configuration to include Terraform CLI and Azure CLI
   - ADMIN and READER service principals creation, test and login
+  - basic Security Management setup using Check Point Terraform module in `management` folder
   
+